refactor: nest queries under dir
