+++ client/TypeScript/src/StateQuery/index.ts
@@ -1,6 +1,2 @@
-export * from './currentEpoch'
-export * from './currentProtocolParameters'
-export * from './eraStart'
-export * from './ledgerTip'
-export * from './proposedProtocolParameters'
+export * from './queries'
 export * from './StateQueryClient'
